+++ gaphor/UML/actions/activitypropertypage.py
@@ -16,7 +16,6 @@ from gaphor.diagram.propertypages import (
 from gaphor.diagram.propertypages import (
     new_builder as diagram_new_builder,
 )
-from gaphor.UML.actions.activity import ActivityItem
 from gaphor.UML.classes.classespropertypages import on_keypress_event
 
 new_builder = new_resource_builder("gaphor.UML.actions")
@@ -25,11 +24,11 @@ new_builder = new_resource_builder("gaphor.UML.actions")
 class ActivityParameters(EditableTreeModel):
     """GTK tree model to edit class attributes."""
 
-    def __init__(self, item):
-        super().__init__(item, cols=(str, object))
+    def __init__(self, subject):
+        super().__init__(subject, cols=(str, object))
 
     def get_rows(self):
-        for node in self._item.subject.node:
+        for node in self._item.node:
             if isinstance(node, UML.ActivityParameterNode):
                 yield [format(node.parameter), node]
 
@@ -37,7 +36,7 @@ class ActivityParameters(EditableTreeModel):
         model = self._item.model
         node = model.create(UML.ActivityParameterNode)
         node.parameter = model.create(UML.Parameter)
-        self._item.subject.node = node
+        self._item.node = node
         return node
 
     @transactional
@@ -51,29 +50,29 @@ class ActivityParameters(EditableTreeModel):
             row[0] = format(node.parameter)
 
     def swap_objects(self, o1, o2):
-        return self._item.subject.node.swap(o1, o2)
+        return self._item.node.swap(o1, o2)
 
     def sync_model(self, new_order):
-        self._item.subject.node.order(
+        self._item.node.order(
             lambda key: new_order.index(key) if key in new_order else -1
         )
 
 
-@PropertyPages.register(ActivityItem)
-class ActivityItemPage(PropertyPageBase):
+@PropertyPages.register(UML.Activity)
+class ActivityPropertyPage(PropertyPageBase):
     order = 40
 
-    def __init__(self, item: ActivityItem):
-        self.item = item
-        self.watcher = item.subject and item.subject.watcher()
+    def __init__(self, subject: UML.Activity):
+        self.subject = subject
+        self.watcher = subject and subject.watcher()
 
     def construct(self):
-        subject = self.item.subject
+        subject = self.subject
 
         if not subject:
             return
 
-        self.model = ActivityParameters(self.item)
+        self.model = ActivityParameters(subject)
 
         builder = new_builder(
             "activity-editor",