+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Allow adding existing proxy port on the owning block's property
 - Fix proxy port drag icon
 - Allow removing proxy port element only if the item from any diagram gets removed
+- Replace ActivityItem specific property page with Activity specific property page
 
 2.21.0
 ------